[SPARKNLP-949] Adding changes for spark 3.5.x backward compatibility
