+++ src/main/scala/com/johnsnowlabs/nlp/AnnotatorModel.scala
@@ -16,11 +16,10 @@
 
 package com.johnsnowlabs.nlp
 
-import com.johnsnowlabs.util.Version
+import com.johnsnowlabs.nlp.util.SparkNlpConfig
 import org.apache.spark.ml.{Model, PipelineModel}
 import org.apache.spark.sql.catalyst.encoders.ExpressionEncoder
 import org.apache.spark.sql.functions._
-import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.{DataFrame, Dataset, Row}
 
 /** This trait implements logic that applies nlp using Spark ML Pipeline transformers Should
@@ -50,7 +49,7 @@ abstract class AnnotatorModel[M <: Model[M]] extends RawAnnotator[M] with CanBeL
         s"${inputAnnotatorTypes.mkString(", ")}")
 
     val inputDataset = beforeAnnotate(dataset)
-
+    val newStructType = inputDataset.schema.add(getOutputCol, Annotation.arrayType)
     val processedDataset = {
       this match {
         case withAnnotate: HasSimpleAnnotate[M] =>
@@ -66,7 +65,8 @@ abstract class AnnotatorModel[M <: Model[M]] extends RawAnnotator[M] with CanBeL
               }
             }))
         case withBatchAnnotate: HasBatchedAnnotate[M] =>
-          implicit val encoder: ExpressionEncoder[Row] = getEncoder(inputDataset)
+          implicit val encoder: ExpressionEncoder[Row] =
+            SparkNlpConfig.getEncoder(inputDataset, newStructType)
           val processedDataFrame = inputDataset.mapPartitions(partition => {
             withBatchAnnotate.batchProcess(partition)
           })
@@ -81,7 +81,8 @@ abstract class AnnotatorModel[M <: Model[M]] extends RawAnnotator[M] with CanBeL
           dfWithMetadata
 
         case withBatchAnnotateImage: HasBatchedAnnotateImage[M] =>
-          implicit val encoder: ExpressionEncoder[Row] = getEncoder(inputDataset)
+          implicit val encoder: ExpressionEncoder[Row] =
+            SparkNlpConfig.getEncoder(inputDataset, newStructType)
           val processedDataFrame = inputDataset.mapPartitions(partition => {
             withBatchAnnotateImage.batchProcess(partition)
           })
@@ -96,7 +97,8 @@ abstract class AnnotatorModel[M <: Model[M]] extends RawAnnotator[M] with CanBeL
           dfWithMetadata
 
         case withBatchAnnotateAudio: HasBatchedAnnotateAudio[M] =>
-          implicit val encoder: ExpressionEncoder[Row] = getEncoder(inputDataset)
+          implicit val encoder: ExpressionEncoder[Row] =
+            SparkNlpConfig.getEncoder(inputDataset, newStructType)
           val processedDataFrame = inputDataset.mapPartitions(partition => {
             withBatchAnnotateAudio.batchProcess(partition)
           })
@@ -113,25 +115,6 @@ abstract class AnnotatorModel[M <: Model[M]] extends RawAnnotator[M] with CanBeL
     }
 
     afterAnnotate(processedDataset)
-  }
-
-  private def getEncoder(inputDataset: Dataset[_]): ExpressionEncoder[Row] = {
-    val newStructType = inputDataset.schema.add(getOutputCol, Annotation.arrayType)
-    val sparkVersion = Version.parse(inputDataset.sparkSession.version).toFloat
-    if (sparkVersion >= 3.5f) {
-      ExpressionEncoder(newStructType)
-    } else {
-      try {
-        // Use reflection to access RowEncoder.apply in older Spark versions
-        val rowEncoderClass = Class.forName("org.apache.spark.sql.catalyst.encoders.RowEncoder")
-        val applyMethod = rowEncoderClass.getMethod("apply", classOf[StructType])
-        applyMethod.invoke(null, newStructType).asInstanceOf[ExpressionEncoder[Row]]
-      } catch {
-        case _: Throwable =>
-          throw new UnsupportedOperationException(
-            "RowEncoder.apply is not supported in this Spark version.")
-      }
-    }
   }
 
   /** Given requirements are met, this applies ML transformation within a Pipeline or stand-alone
