+++ src/main/scala/com/johnsnowlabs/nlp/annotators/common/DatasetHelpers.scala
@@ -18,8 +18,8 @@ package com.johnsnowlabs.nlp.annotators.common
 
 import com.johnsnowlabs.ml.crf.TextSentenceLabels
 import com.johnsnowlabs.ml.tensorflow.SentenceGrouper
+import com.johnsnowlabs.nlp.util.SparkNlpConfig
 import org.apache.spark.sql.DataFrame
-import org.apache.spark.sql.catalyst.encoders.ExpressionEncoder
 
 import scala.reflect.ClassTag
 
@@ -27,7 +27,7 @@ object DatasetHelpers {
 
   implicit class DataFrameHelper(dataset: DataFrame) {
     def randomize: DataFrame = {
-      implicit val encoder = ExpressionEncoder(dataset.schema)
+      implicit val encoder = SparkNlpConfig.getEncoder(dataset, dataset.schema)
       dataset.mapPartitions {
         new scala.util.Random().shuffle(_).toIterator
       }
